pass freq to _shift
